Add files via upload
Made adjustments to the skeleton of the code. Everything seems to be in working order. Now we just need to write in all of the functions to derive the features. After this, we can generate 20 different files for each of the combinations of Hz and Window so that ML techniques can be applied to them.